# taskmanager/generated-sources/src/main/java/org/openapitools/api/TasksApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (6.6.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.openapitools.api;

import org.openapitools.model.TasksGet200ResponseInner;
import org.openapitools.model.TasksPostRequest;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.annotation.Generated;
import jakarta.validation.Valid;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import java.util.List;
import java.util.Map;
import java.util.Optional;


@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-03-31T03:08:15.250988-03:00[America/Santiago]")
@Validated
@Tag(name = "tasks", description = "the tasks API")
public interface TasksApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * GET /tasks : Get all tasks
     *
     * @return List of tasks (status code 200)
     */
    @Operation(
        operationId = "tasksGet",
        summary = "Get all tasks",
        responses = {
            @ApiResponse(responseCode = "200", description = "List of tasks", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = TasksGet200ResponseInner.class)))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/tasks",
        produces = { "application/json" }
    )
    default ResponseEntity<List<TasksGet200ResponseInner>> tasksGet(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"description\" : \"description\", \"id\" : 0, \"title\" : \"title\", \"status\" : \"status\" }, { \"description\" : \"description\", \"id\" : 0, \"title\" : \"title\", \"status\" : \"status\" } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * DELETE /tasks/{id} : Delete a task
     *
     * @param id  (required)
     * @return Task deleted (status code 204)
     */
    @Operation(
        operationId = "tasksIdDelete",
        summary = "Delete a task",
        responses = {
            @ApiResponse(responseCode = "204", description = "Task deleted")
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/tasks/{id}"
    )
    default ResponseEntity<Void> tasksIdDelete(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Integer id
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * PUT /tasks/{id} : Update a task
     *
     * @param id  (required)
     * @param tasksPostRequest  (required)
     * @return Task updated (status code 200)
     */
    @Operation(
        operationId = "tasksIdPut",
        summary = "Update a task",
        responses = {
            @ApiResponse(responseCode = "200", description = "Task updated")
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/tasks/{id}",
        consumes = { "application/json" }
    )
    default ResponseEntity<Void> tasksIdPut(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Integer id,
        @Parameter(name = "TasksPostRequest", description = "", required = true) @Valid @RequestBody TasksPostRequest tasksPostRequest
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /tasks : Create a new task
     *
     * @param tasksPostRequest  (required)
     * @return Task created (status code 201)
     */
    @Operation(
        operationId = "tasksPost",
        summary = "Create a new task",
        responses = {
            @ApiResponse(responseCode = "201", description = "Task created")
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/tasks",
        consumes = { "application/json" }
    )
    default ResponseEntity<Void> tasksPost(
        @Parameter(name = "TasksPostRequest", description = "", required = true) @Valid @RequestBody TasksPostRequest tasksPostRequest
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}