# taskmanager/src/main/java/com/example/taskmanager/model/RegisterRequest.java
package com.example.taskmanager.model;

public class RegisterRequest {
    public String getUsername() {
		return username;
	}
	public void setUsername(String username) {
		this.username = username;
	}
	public String getPassword() {
		return password;
	}
	public void setPassword(String password) {
		this.password = password;
	}
	public String getEmail() {
		return email;
	}
	public void setEmail(String email) {
		this.email = email;
	}
	private String username;
    private String password;
    private String email;

    
}